requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==3.2.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyrsistent==0.20.0
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.3
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
gunicorn==20.1.0